+++ app/controllers/user_files_controller.rb
@@ -1,44 +1,64 @@
 class UserFilesController < ApplicationController
   def index
-    @user_files = current_user.user_files
-    @user_file = UserFile.new
-    render :action => "new"
+    if not logged_in? then
+      redirect_to controller: "sessions", action: "new"
+    else
+      @user_files = current_user.user_files
+      @user_file = UserFile.new
+      render :action => "new"
+    end
   end
 
   def new
-    @user_file = UserFile.new
-    @user_files = current_user.user_files
+    if not logged_in? then
+      redirect_to controller: "sessions", action: "new"
+    else
+      @user_file = UserFile.new
+      @user_files = current_user.user_files
+    end
   end
 
   def create_for_jobform
-    respond_to do |format|
-      @user_file = UserFile.new(file_params)
-      @user_file.user = current_user
-      if @user_file.save then
-        @user_file.file.canonicalize_seq!
+    if not logged_in? then
+      redirect_to controller: "sessions", action: "new"
+    else
+      respond_to do |format|
+        @user_file = UserFile.new(file_params)
+        @user_file.user = current_user
+        if @user_file.save then
+          @user_file.file.canonicalize_seq!
+        end
+        format.js { render 'jobs/user_file_line'}
       end
-      format.js { render 'jobs/user_file_line'}
     end
   end
 
   def create
-    respond_to do |format|
-      @user_file = UserFile.new(file_params)
-      @user_file.user = current_user
-      if @user_file.save then
-        @user_file.file.canonicalize_seq!
+    if not logged_in? then
+      redirect_to controller: "sessions", action: "new"
+    else
+      respond_to do |format|
+        @user_file = UserFile.new(file_params)
+        @user_file.user = current_user
+        if @user_file.save then
+          @user_file.file.canonicalize_seq!
+        end
+        format.html { redirect_to new_user_file_path }
+        format.js
       end
-      format.html { redirect_to new_user_file_path }
-      format.js
     end
   end
 
   def destroy
-    thisfile = UserFile.find(params[:id])
-    if thisfile then
-      thisfile.destroy
+    if not logged_in? then
+      redirect_to controller: "sessions", action: "new"
+    else
+      thisfile = UserFile.find(params[:id])
+      if thisfile then
+        thisfile.destroy
+      end
+      redirect_to action: "new"
     end
-    redirect_to action: "new"
   end
 
   private
